requirements.txt
Django==1.5.1
MySQL-python==1.2.5
PyYAML==3.10
South==1.0.0
django-crispy-forms==1.4.0
django-debug-toolbar==0.9.4
pillow==2.5.1
python-dateutil==2.1
virtualenv==1.8.4
gunicorn==0.17.4
xlwt==0.7.5
flup==1.0